workers/requirements.txt
httpx>=0.27,<0.29
openai>=1.40,<2
anthropic>=0.34,<1
pyyaml>=6.0.1,<7
psycopg[binary]>=3.1,<4
